+++ lib/src/caches/ephemeral_fifo_cache.dart
@@ -58,6 +58,14 @@ class EphemeralFIFOCache<K, V> extends ThreadSafeCache<K, V> {
     });
   }
 
+  /// Retrieves [key] without removing it from the cache.
+  ///
+  /// **This method is async-safe.**
+  @override
+  Future<V?> peek(K key) async {
+    return await _lock.synchronized(() => _cache[key]);
+  }
+
   /// Checks whether [key] exists in the cache without removing it.
   ///
   /// **This method is async-safe.**